Refactored resource_center/tests.py into funtional_tests package
